Add README.md

+++ controllers/catController.js
@@ -16,15 +16,15 @@ const storage = multer.diskStorage({
 
 const getAllCats = async (req, res) => {
     try {
-        const cats = await Cat.find()
-        res.render('home', {cats: cats}) //render home page, pass cat object with all cat data back. EJS reads cat data and renders on page
+        const cats = await Cat.find().populate('owner')
+        res.render('home', {cats: cats, user: req.user}) //render home page, pass cat object with all cat data back. EJS reads cat data and renders on page
     } catch (error) {
         console.log(error)
     }
 }
 
 const uploadPage = (req, res) => {
-    res.render('upload')
+    res.render('upload', {user: req.user})
 }
 
 //create cat, Cat is uppercase as it accesses the Cat model that is imported at the beginning of the page
@@ -35,7 +35,8 @@ const createCat = async (req, res) => {
             age: req.body.age,
             favoriteFood: req.body.favoriteFood,
             funFact: req.body.funFact,
-            image: req.file.filename
+            image: req.file.filename,
+            owner: req.user._id
     })
 
     await cat.save()
@@ -49,7 +50,7 @@ const createCat = async (req, res) => {
 const editPage = async (req, res) => {
     try {
       const cat = await Cat.findById(req.params.id);
-      res.render('edit', { cat: cat });
+      res.render('edit', { cat: cat, user: req.user });
     } catch(err) {
       console.log(err);
     }
@@ -59,7 +60,10 @@ const updateCat = async (req, res) => {
     console.log(req.params.id)
     console.log(req.body)
     try {
+      let cat = await Cat.findById(req.params.id)
+      if(cat.owner.equals(req.user.id)) {
       await Cat.findByIdAndUpdate(req.params.id, req.body);
+      }
       res.redirect('/');
     } catch(err) {
       console.log(err);
@@ -70,7 +74,10 @@ const updateCat = async (req, res) => {
     console.log(req.params.id)
     console.log(req.body)
     try {
+      let cat = await Cat.findById(req.params.id)
+      if(cat.owner.equals(req.user.id)) {
       await Cat.findByIdAndRemove(req.params.id);
+      }
       res.redirect('/');
     } catch(err) {
       console.log(err);
